move endpoints to correct config
