--- chessdotcom/_endpoints/__init__.py
@@ -1,1 +1,0 @@
-from .get_player_profile import get_player_profile
