+++ tox.ini
@@ -5,7 +5,7 @@ extend-ignore = E203
 per-file-ignores =
     chessdotcom/aio.py: F401,F403
     __init__.py: F401,E402
-    chessdotcom/endpoints.py: F401
+    chessdotcom/endpoints/__init__.py: F401
     
 [isort]
 profile = black
